# Closes #6460 Maybe optimize lookups of tables in DB to stop using SHOW TABLES LIKE