--- a/inc/Engine/Common/Database/Queries/AbstractQuery.php
+++ b/inc/Engine/Common/Database/Queries/AbstractQuery.php
@@ -520,9 +520,8 @@ class AbstractQuery extends Query {
 		}
 
 		// Query statement.
-		$query    = 'SHOW TABLES LIKE %s';
-		$like     = $db->esc_like( $db->{$this->table_name} );
-		$prepared = $db->prepare( $query, $like );
+		$query    = 'SELECT table_name FROM information_schema.tables WHERE table_name = %s LIMIT 1';
+		$prepared = $db->prepare( $query, $this->table_name );
 		$result   = $db->get_var( $prepared );
 
 		// Does the table exist?