--- a/inc/Engine/Common/Database/Queries/AbstractQuery.php
+++ b/inc/Engine/Common/Database/Queries/AbstractQuery.php
@@ -521,7 +521,7 @@ class AbstractQuery extends Query {
 
 		// Query statement.
 		$query    = 'SELECT table_name FROM information_schema.tables WHERE table_name = %s LIMIT 1';
-		$prepared = $db->prepare( $query, $this->table_name );
+		$prepared = $db->prepare( $query, $db->{$this->table_name} );
 		$result   = $db->get_var( $prepared );
 
 		// Does the table exist?