--- a/inc/Engine/Common/PerformanceHints/Database/Queries/AbstractQueries.php
+++ b/inc/Engine/Common/PerformanceHints/Database/Queries/AbstractQueries.php
@@ -121,9 +121,8 @@ class AbstractQueries extends Query {
 		}
 
 		// Query statement.
-		$query    = 'SHOW TABLES LIKE %s';
-		$like     = $db->esc_like( $db->{$this->table_name} );
-		$prepared = $db->prepare( $query, $like );
+		$query    = 'SELECT table_name FROM information_schema.tables WHERE table_name = %s LIMIT 1';
+		$prepared = $db->prepare( $query, $this->table_name );
 		$result   = $db->get_var( $prepared );
 
 		// Does the table exist?