# Description
Changed lookup table query.

Fixes #6460 


## Type of change

- [x] Enhancement (non-breaking change which improves an existing functionality).

## Detailed scenario
With a large site(mostly multisites), the query is run on every page loads and this slows down the page.
More information here https://wp-media.slack.com/archives/C43T1AYMQ/p1709561232039849

## Technical description
Changed the query from `show tables` to select with a limit of 1

### Documentation


# Mandatory Checklist

## Code validation

- [x] I validated all the Acceptance Criteria. If possible, provide screenshots or videos.
- [x] I triggered all changed lines of code at least once without new errors/warnings/notices.

## Code style
- [x] I wrote a self-explanatory code about what it does.
- [x] I protected entry points against unexpected inputs.
- [x] I did not introduce unnecessary complexity.