@MathieuLamiot 2024-09-17
@wp-media/qa-team @Khadreal to discuss how to validate / test this.
A quick feedback here: the main risk I see is that we are not using "LIKE" anymore but a strictly equal table name. We would need to check on different setups (multi-site, a few hosting, etc.) that we always have the exact table name.
@Khadreal for instance, in most queries, we use $db->prefix, which does not seem to be the case in your PR. We'd need to ensure we are not missing a prefix mechanism here.
https://github.com/wp-media/wp-rocket/blob/b898c2a8b6f52d9fd93702a59419d57f208b0b4b/inc/Engine/Preload/Database/Queries/Cache.php#L310-L311