@Mai-Saad 2024-09-19
Test plan:
- confirm fixing the performance issue
- NRT plan execution https://wpmediaqa.testrail.io/index.php?/runs/view/898&group_by=cases:section_id&group_order=asc