@Khadreal 2024-09-19
> @wp-media/qa-team @Khadreal to discuss how to validate / test this. A quick feedback here: the main risk I see is that we are not using "LIKE" anymore but a strictly equal table name. We would need to check on different setups (multi-site, a few hosting, etc.) that we always have the exact table name. @Khadreal for instance, in most queries, we use $db->prefix, which does not seem to be the case in your PR. We'd need to ensure we are not missing a prefix mechanism here.
> 
> https://github.com/wp-media/wp-rocket/blob/b898c2a8b6f52d9fd93702a59419d57f208b0b4b/inc/Engine/Preload/Database/Queries/Cache.php#L310-L311

@MathieuLamiot 
Tested this on multisite and its work as it should. `$db->{$this->table_name}`  will return the table name with prefix. This `$this->get_db()` returns the db name with prefix https://github.com/wp-media/wp-rocket/blob/e951e59168ca0b3b3b1c378a1c2af70a883588bb/inc/Engine/Common/Database/Queries/AbstractQuery.php#L234