@Miraeld 2024-09-19
Hello hello, 

I've conducted some tests and some logs of execution time between the old version and new one. 

- For a database with 100 tables, it decreased from an average of 15ms to 5ms, representing a 66% improvement in performance.
- When I scaled up to a database with 500 tables, the improvement was even more significant. The execution time decreased from an average of 75ms to 15ms, which is an 80% improvement.
- Finally, in a large-scale scenario with a database containing 100,000 tables, the execution time went down from an average of 2 seconds to 20-30 ms. 

So I confirm that it works as expected.